Add support for webauthn identity into many
