+++ src/many-identity-dsa/src/impls/ecdsa.rs
@@ -72,6 +72,12 @@ pub fn public_key(key: &CoseKey) -> Result<Option<CoseKey>, ManyError> {
     }
 }
 
+/// Extract the address of a CoseKey, if it implements ECDSA.
+pub fn address(key: &CoseKey) -> Result<Address, ManyError> {
+    let public_key = public_key(&key)?.ok_or_else(|| ManyError::unknown("Could not load key."))?;
+    unsafe { cose::address_unchecked(&public_key) }
+}
+
 #[derive(Clone, Debug)]
 struct EcDsaIdentityInner {
     address: Address,
