+++ src/many-identity-webauthn/Cargo.toml
@@ -11,13 +11,28 @@ coset = "0.3.2"
 many-error = { path = "../many-error" }
 many-identity = { path = "../many-identity" }
 many-identity-dsa = { path = "../many-identity-dsa", features = ["ecdsa"] }
+many-modules = { path = "../many-modules", optional = true }
 many-protocol = { path = "../many-protocol" }
 minicbor = "0.18.0"
 once_cell = "1.13.0"
+rand = { version = "0.8.5", optional = true }
 serde = "1.0.142"
 serde_json = "1.0.83"
 sha2 = "0.10.2"
 tracing = "0.1.36"
+webauthn-authenticator-rs = { version = "0.4.9", optional = true, features = ["u2fhid"] }
+webauthn-rs = { version = "0.4.8", optional = true }
+webauthn-rs-proto = { version = "0.4.9", optional = true }
 
 [dev-dependencies]
 hex = "0.4.3"
+
+[features]
+default = ["identity"]
+identity = [
+    "many-modules",
+    "rand",
+    "webauthn-authenticator-rs",
+    "webauthn-rs",
+    "webauthn-rs-proto",
+]
