+++ src/many/src/main.rs
@@ -8,12 +8,13 @@ use many_identity::verifiers::AnonymousVerifier;
 use many_identity::{Address, AnonymousIdentity, Identity};
 use many_identity_dsa::{CoseKeyIdentity, CoseKeyVerifier};
 use many_identity_hsm::{Hsm, HsmIdentity, HsmMechanismType, HsmSessionType, HsmUserType};
+use many_identity_webauthn::WebAuthnIdentity;
 use many_mock::{parse_mockfile, server::ManyMockServer, MockEntries};
-use many_modules::ledger;
 use many_modules::r#async::attributes::AsyncAttribute;
 use many_modules::r#async::{StatusArgs, StatusReturn};
+use many_modules::{idstore, ledger};
 use many_protocol::{
-    encode_cose_sign1_from_request, RequestMessage, RequestMessageBuilder, ResponseMessage,
+    encode_cose_sign1_from_request, ManyUrl, RequestMessage, RequestMessageBuilder, ResponseMessage,
 };
 use many_server::transport::http::HttpServer;
 use many_server::ManyServer;
@@ -50,6 +51,9 @@ enum SubCommand {
 
     /// Display the textual ID of a public key located on an HSM.
     HsmId(HsmIdOpt),
+
+    /// Display the textual ID of a webauthn key.
+    WebauthnId(WebauthnIdOpt),
 
     /// Creates a message and output it.
     Message(MessageOpt),
@@ -88,6 +92,21 @@ struct HsmIdOpt {
 }
 
 #[derive(Parser)]
+struct WebauthnIdOpt {
+    /// URL to the relying party (the MANY server implementing idstore).
+    rp: ManyUrl,
+
+    /// The recall phrase.
+    #[clap(long, conflicts_with("address"))]
+    phrase: Option<String>,
+
+    /// The address of the webauthn key. This may seem redundant but in this
+    /// case the webauthn flow will still be checked to get the ID.
+    #[clap(long, conflicts_with("phrase"))]
+    address: Option<Address>,
+}
+
+#[derive(Parser)]
 #[clap(
     group(
         ArgGroup::new("hsm")
@@ -105,6 +124,27 @@ struct MessageOpt {
     /// A pem file to sign the message. If this is omitted, the message will be anonymous.
     #[clap(long)]
     pem: Option<PathBuf>,
+
+    /// Use Webauthn as the authentication scheme.
+    #[clap(long, conflicts_with("pem"))]
+    webauthn: bool,
+
+    /// The origin to use in the webauthn flow. By default will use the
+    /// relying party's protocol, hostname and port.
+    #[clap(long, requires("webauthn"))]
+    webauthn_origin: Option<ManyUrl>,
+
+    /// The Webauthn provider. By default will use the same server.
+    #[clap(long, requires("webauthn"))]
+    rp: Option<ManyUrl>,
+
+    /// The recall phrase for webauthn.
+    #[clap(long, requires("webauthn"), conflicts_with("address"))]
+    phrase: Option<String>,
+
+    /// The address for webauthn.
+    #[clap(long, requires("webauthn"), conflicts_with("phrase"))]
+    address: Option<Address>,
 
     /// Timestamp (in seconds since epoch).
     #[clap(long)]
@@ -315,6 +355,44 @@ async fn message_from_hex(
     show_response(&response, client, r#async).await
 }
 
+async fn create_webauthn_identity(
+    rp: ManyUrl,
+    origin: Option<ManyUrl>,
+    phrase: Option<String>,
+    address: Option<Address>,
+) -> WebAuthnIdentity {
+    let client = ManyClient::new(rp.clone(), Address::anonymous(), AnonymousIdentity)
+        .expect("Could not create client");
+
+    let response = if let Some(phrase) = phrase {
+        client
+            .call(
+                "idstore.getFromRecallPhrase",
+                idstore::GetFromRecallPhraseArgs(phrase.split(" ").map(String::from).collect()),
+            )
+            .await
+            .unwrap()
+    } else if let Some(address) = address {
+        client
+            .call(
+                "idstore.getFromAddress",
+                idstore::GetFromAddressArgs(address),
+            )
+            .await
+            .unwrap()
+    } else {
+        error!("Must specify a phrase or address.");
+        process::exit(3);
+    };
+
+    let get_returns = response.data.expect("Error from the server");
+    let get_returns =
+        minicbor::decode::<idstore::GetReturns>(&get_returns).expect("Deserialization error");
+
+    WebAuthnIdentity::authenticate(origin.unwrap_or(rp), get_returns)
+        .expect("Could not create Identity object")
+}
+
 #[tokio::main]
 async fn main() {
     let Opts {
@@ -370,7 +448,7 @@ async fn main() {
                 println!("{i}");
             } else {
                 error!("Could not understand the argument.");
-                std::process::exit(2);
+                process::exit(2);
             }
         }
         SubCommand::HsmId(o) => {
@@ -397,6 +475,10 @@ async fn main() {
             }
 
             println!("{id}");
+        }
+        SubCommand::WebauthnId(o) => {
+            let identity = create_webauthn_identity(o.rp, None, o.phrase, o.address).await;
+            println!("{}", identity.address());
         }
         SubCommand::Message(o) => {
             let to_identity = o.to.unwrap_or_default();
@@ -436,6 +518,15 @@ async fn main() {
             } else if let Some(p) = o.pem {
                 // If `pem` is not provided, use anonymous and don't sign.
                 Box::new(CoseKeyIdentity::from_pem(std::fs::read_to_string(p).unwrap()).unwrap())
+            } else if o.webauthn {
+                let rp =
+                    o.rp.as_ref()
+                        .or(o.server.as_ref())
+                        .expect("Must pass a server or --rp");
+                let identity =
+                    create_webauthn_identity(rp.clone(), o.webauthn_origin, o.phrase, o.address)
+                        .await;
+                Box::new(identity)
             } else {
                 Box::new(AnonymousIdentity)
             };
